update module names, fix icon rendering issue causing test problems (rabbithole)
